requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
discord.py==1.5.0
idna==2.10
isort==5.6.1
lazy-object-proxy==1.5.1
mccabe==0.6.1
multidict==4.7.6
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wavelink==0.9.6
websockets==8.1
wrapt==1.12.1
yarl==1.6.0